add image preview and form wrapping
